refactor(registration): split monolithic registration.py into package
BREAKING CHANGE: Internal import path changed from jaxarc.registration
(single file) to jaxarc.registration (package). Public API unchanged —
all symbols re-exported from __init__.py.
